mix(scm&structure): 1. add Fossil scm fix #35 2. move golang/go source to internal dir
