+++ Dockerfile
@@ -1,15 +1,17 @@
 FROM golang:1.11 AS build
 
-env GO111MODULE on
-env GOROOT /usr/local/go
-
 COPY ./ /goproxy
-WORKDIR /goproxy
-RUN go generate
-RUN go build
 
-FROM buildpack-deps:stretch-scm
+RUN cd /goproxy &&\
+    export GO111MODULE=on &&\
+    go generate &&\
+    go mod tidy &&\
+    go build
+
+FROM alpine:3.8
+RUN apk add --no-cache git mercurial subversion bzr fossil
 COPY --from=build /goproxy/goproxy /bin/goproxy
+
 EXPOSE 8081
 
 CMD ["goproxy"]
